requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.11.28
Django==2.2.5
gast==0.2.2
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.9.0
imageio==2.6.1
ISR==2.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.2
pyaml==19.12.0
pyreadline==2.1
pytz==2019.3
PyYAML==5.3
scipy==1.3.2
six==1.13.0
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
tqdm==4.41.1
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
